created modular message components and message hook
