+++ src/components/Dashboard/ActiveGame/GameChat/MessageContent/MessageContentTypes.ts
@@ -1,0 +1,8 @@
+import { Message } from "../GameChatTypes";
+import { RefObject } from "react";
+
+export interface MessageContentProps {
+  messages: Message[];
+  retrySendMessage: () => void;
+  message_txt_container: RefObject<HTMLUListElement>;
+}
