+++ src/components/Dashboard/ActiveGame/GameChat/MessageIcon/MessageIcon.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import Badge from '@mui/material/Badge';
+import { MessageIconProps } from "./MessageIconTypes";
+
+export const MessageIcon = ({ 
+    messageBadge, 
+    messageStyle, 
+    showMessages 
+}: MessageIconProps) => {
+  return (
+    <div className="message-icon-container">
+      <Badge onClick={showMessages} badgeContent={messageBadge} color="primary">
+        <img className="message-icon" src={messageStyle} alt="Message Icon" />
+      </Badge>
+    </div>
+  );
+};
