+++ src/components/Dashboard/ActiveGame/GameChat/MessageIcon/MessageIconTypes.ts
@@ -1,0 +1,6 @@
+
+export interface MessageIconProps {
+    messageBadge: number;
+    messageStyle: string;
+    showMessages: () => void;
+}
